WIP: better color picker library

+++ .github/build-utils.mjs
@@ -2,17 +2,24 @@
 
 import { promises as fs } from 'fs';
 import path from 'path';
-
-const initialDir = process.cwd(); // Save the initial directory
-
+const args = process.argv.slice(2);
+const skipBuild = args.includes('--skip-build');
+// script dir
+const __dirname = path.dirname(new URL(import.meta.url).pathname);
+// the dir at the top of the script dir
+const rootDir = path.resolve(__dirname, '../../');
+const initialDir = rootDir;
+cd(rootDir);
 (async () => {
-    try {
-        cd('wp-ultimate-tables/new-frontend'); // Change to the directory
-        await $`npm run build`;
-        console.log('Command completed successfully');
-        cd(initialDir); // Change back to the initial directory
-    } catch (error) {
-        console.error('Command failed with status:', error.exitCode);
+    if ( !skipBuild ) {
+        try {
+            cd('wp-ultimate-tables/new-frontend'); // Change to the directory
+            await $`npm run build`;
+            console.log('Command completed successfully');
+            cd(initialDir); // Change back to the initial directory
+        } catch (error) {
+            console.error('Command failed with status:', error.exitCode);
+        }
     }
     const dir = 'wp-ultimate-tables/new-frontend/dist';
     const files = await fs.readdir(dir);
@@ -22,19 +29,23 @@ const initialDir = process.cwd(); // Save the initial directory
         await $`cp ${dir}/${jsFile} wp-ultimate-tables/assets/js/frontend.js`;
         console.log('File copied successfully');
     } else {
-        console.log('No .js file found in the directory');
+        console.log('FRONTEND: No .js file found in the directory');
     }
-    // remove all the files in the directory
-    await $`rm -rf wp-ultimate-tables/new-frontend/dist/*`;
+    if ( !skipBuild ) {
+        // remove all the files in the directory
+        await $`rm -rf wp-ultimate-tables/new-frontend/dist/*`;
+    }
 })();
 (async () => {
-    try {
-        cd('wp-ultimate-tables/webpack-backend'); // Change to the directory
-        await $`npm run build`;
-        console.log('Command completed successfully');
-        cd(initialDir); // Change back to the initial directory
-    } catch (error) {
-        console.error('Command failed with status:', error.exitCode);
+    if ( !skipBuild ) {
+        try {
+            cd('wp-ultimate-tables/webpack-backend'); // Change to the directory
+            await $`npm run build`;
+            console.log('Command completed successfully');
+            cd(initialDir); // Change back to the initial directory
+        } catch (error) {
+            console.error('Command failed with status:', error.exitCode);
+        }
     }
     const dir = initialDir + '/wp-ultimate-tables/webpack-backend/dist';
     // const files = await fs.readdir(dir);
@@ -44,8 +55,10 @@ const initialDir = process.cwd(); // Save the initial directory
         await $`cp ${dir}/${jsFile} ${initialDir}/wp-ultimate-tables/assets/js/admin.js`;
         console.log('File copied successfully');
     } else {
-        console.log('No .js file found in the directory');
+        console.log('BACKEND : No .js file found in the directory');
     }
-    // remove all the files in the directory
-    await $`rm -rf wp-ultimate-tables/webpack-backend/dist/*`;
+    if ( !skipBuild ) {
+        // remove all the files in the directory
+        await $`rm -rf wp-ultimate-tables/webpack-backend/dist/*`;
+    }
 })();
